updated my solution to include the check that I removed
the algorithm now checks if the place has already been visited with the
same direction. this makes sure that it does not stay in place running
in circles, while the rest of the algo goes on.

also added a better comparison. although it is still kind of a mystery
why the __eq__ method does not work...