added a benchmarking.py file for benchmarking the 2048 game
I made this for the people who are not using linux or wsl, since it
works crossplatform.

You can use --input to give the file and -n to give the number of
iterations

You also need for the puzzle.py to print the score to stdout, so that it
can capture it. Something like `print(f"score is: {score}")`. the :
character is required for the parsing to work.