added that docker compose file that I used to self host
this uses cloudflare tunnels, so that you can host it on your own
machine and publish it to the internet. you can change some traefik
values to make it fit vps use instead as well.

it also uses docker image, so you dont need to build or clone the
project itself.